+++ R/s4-AlphaStableBernsteinFunction.R
@@ -40,7 +40,7 @@
 #'   \insertAllCited{}
 #'
 #' @seealso [levyDensity()], [stieltjesDensity()], [valueOf()],
-#'   [intensities()], [calcExShockArrivalIntensities()],
+#'   [calcShockArrivalIntensities()], [calcExShockArrivalIntensities()],
 #'   [calcExShockSizeArrivalIntensities()], [calcMDCMGeneratorMatrix()],
 #'   [rextmo()], [rpextmo()]
 #'
@@ -83,9 +83,9 @@
 #'
 #' # Calculate shock-arrival intensities
 #' bf <- AlphaStableBernsteinFunction(alpha = 0.8)
-#' intensities(bf, 3)
-#' intensities(bf, 3, method = "stieltjes")
-#' intensities(bf, 3, tolerance = 1e-4)
+#' calcShockArrivalIntensities(bf, 3)
+#' calcShockArrivalIntensities(bf, 3, method = "stieltjes")
+#' calcShockArrivalIntensities(bf, 3, tolerance = 1e-4)
 #'
 #' # Calculate exchangeable shock-arrival intensities
 #' bf <- AlphaStableBernsteinFunction(alpha = 0.4)