+++ R/s4-ExponentialBernsteinFunction.R
@@ -34,9 +34,10 @@
 #' @references
 #'   \insertAllCited{}
 #'
-#' @seealso [levyDensity()], [stieltjesDensity()], [valueOf()], [intensities()],
-#'   [calcExShockArrivalIntensities()], [calcExShockSizeArrivalIntensities()],
-#'   [calcMDCMGeneratorMatrix()], [rextmo()], [rpextmo()]
+#' @seealso [levyDensity()], [stieltjesDensity()], [valueOf()],
+#'   [calcShockArrivalIntensities()], [calcExShockArrivalIntensities()],
+#'   [calcExShockSizeArrivalIntensities()], [calcMDCMGeneratorMatrix()],
+#'   [rextmo()], [rpextmo()]
 #'
 #' @docType class
 #' @name ExponentialBernsteinFunction-class
@@ -73,9 +74,9 @@
 #'
 #' # Calculate shock-arrival intensities
 #' bf <- ExponentialBernsteinFunction(lambda = 0.8)
-#' intensities(bf, 3)
-#' intensities(bf, 3, method = "levy")
-#' intensities(bf, 3, tolerance = 1e-4)
+#' calcShockArrivalIntensities(bf, 3)
+#' calcShockArrivalIntensities(bf, 3, method = "levy")
+#' calcShockArrivalIntensities(bf, 3, tolerance = 1e-4)
 #'
 #' # Calculate exchangeable shock-arrival intensities
 #' bf <- ExponentialBernsteinFunction(lambda = 0.4)