+++ R/s4-InverseGaussianBernsteinFunction.R
@@ -50,9 +50,10 @@
 #' @references
 #'  \insertAllCited{}
 #'
-#' @seealso [levyDensity()], [stieltjesDensity()], [valueOf()], [intensities()],
-#'   [calcExShockArrivalIntensities()], [calcExShockSizeArrivalIntensities()],
-#'   [calcMDCMGeneratorMatrix()], [rextmo()], [rpextmo()]
+#' @seealso [levyDensity()], [stieltjesDensity()], [valueOf()],
+#'   [calcShockArrivalIntensities()], [calcExShockArrivalIntensities()],
+#'   [calcExShockSizeArrivalIntensities()], [calcMDCMGeneratorMatrix()],
+#'   [rextmo()], [rpextmo()]
 #'
 #' @docType class
 #' @name InverseGaussianBernsteinFunction-class
@@ -93,9 +94,9 @@
 #'
 #' # Calculate shock-arrival intensities
 #' bf <- InverseGaussianBernsteinFunction(eta = 0.8)
-#' intensities(bf, 3)
-#' intensities(bf, 3, method = "stieltjes")
-#' intensities(bf, 3, tolerance = 1e-4)
+#' calcShockArrivalIntensities(bf, 3)
+#' calcShockArrivalIntensities(bf, 3, method = "stieltjes")
+#' calcShockArrivalIntensities(bf, 3, tolerance = 1e-4)
 #'
 #' # Calculate exchangeable shock-arrival intensities
 #' bf <- InverseGaussianBernsteinFunction(eta = 0.4)