+++ tests/testthat/helper-parameter.R
@@ -5,7 +5,7 @@ calc_theta_constant <- function(d, constant) {
 
 intensities_constant <- function(d, constant) {
   bf <- ConstantBernsteinFunction(constant = constant)
-  intensities(bf, d)
+  calcShockArrivalIntensities(bf, d)
 }
 
 calc_theta_linear <- function(d, scale) {
@@ -15,7 +15,7 @@ calc_theta_linear <- function(d, scale) {
 
 intensities_linear <- function(d, scale) {
   bf <- LinearBernsteinFunction(scale = scale)
-  intensities(bf, d)
+  calcShockArrivalIntensities(bf, d)
 }
 
 calc_theta_armageddon <- function(d, alpha, beta) {
@@ -35,7 +35,7 @@ calc_theta_poisson <- function(d, eta) {
 
 intensities_poisson <- function(d, eta) {
   bf <- PoissonBernsteinFunction(eta = eta)
-  intensities(bf, d)
+  calcShockArrivalIntensities(bf, d)
 }
 
 calc_theta_alpha_stable <- function(d, alpha) {
@@ -45,7 +45,7 @@ calc_theta_alpha_stable <- function(d, alpha) {
 
 intensities_alpha_stable <- function(d, alpha) {
   bf <- AlphaStableBernsteinFunction(alpha = alpha)
-  intensities(bf, d)
+  calcShockArrivalIntensities(bf, d)
 }
 
 calc_theta_exponential <- function(d, lambda) {
@@ -55,7 +55,7 @@ calc_theta_exponential <- function(d, lambda) {
 
 intensities_exponential <- function(d, lambda) {
   bf <- ExponentialBernsteinFunction(lambda = lambda)
-  intensities(bf, d)
+  calcShockArrivalIntensities(bf, d)
 }
 
 calc_theta_gamma <- function(d, a) {
@@ -65,7 +65,7 @@ calc_theta_gamma <- function(d, a) {
 
 intensities_gamma <- function(d, a) {
   bf <- GammaBernsteinFunction(a = a)
-  intensities(bf, d)
+  calcShockArrivalIntensities(bf, d)
 }
 
 calc_theta_pareto <- function(d, alpha, x0) {
@@ -75,7 +75,7 @@ calc_theta_pareto <- function(d, alpha, x0) {
 
 intensities_pareto <- function(d, alpha, x0) {
   bf <- ParetoBernsteinFunction(alpha = alpha, x0 = x0)
-  intensities(bf, d)
+  calcShockArrivalIntensities(bf, d)
 }
 
 calc_theta_inverse_gaussian <- function(d, eta) { # nolint
@@ -85,15 +85,15 @@ calc_theta_inverse_gaussian <- function(d, eta) { # nolint
 
 intensities_inverse_gaussian <- function(d, eta) {
   bf <- InverseGaussianBernsteinFunction(eta = eta)
-  intensities(bf, d)
+  calcShockArrivalIntensities(bf, d)
 }
 
-intensities_hierarchical <- function(d1, d2, lambda, eta, a, alpha) { # nolint
+intensities_hierarchical <- function(d1, d2, gamma, eta, a, alpha) { # nolint
   theta_1 <- calc_theta_gamma(d1, a)
   theta_2 <- calc_theta_alpha_stable(d2, alpha)
 
-  intensities <- lambda * intensities_poisson(d1 + d2, eta)
-  for (j in seq_along(intensities)) {
+  lambda <- gamma * intensities_poisson(d1 + d2, eta)
+  for (j in seq_along(lambda)) {
     count_1 <- 0
     count_2 <- 0
     for (i in 1:d1) {
@@ -104,11 +104,11 @@ intensities_hierarchical <- function(d1, d2, lambda, eta, a, alpha) { # nolint
     }
 
     if (count_1 > 0 && count_2 == 0) {
-      intensities[j] <- intensities[j] + theta_1[count_1]
+      lambda[j] <- lambda[j] + theta_1[count_1]
     } else if (count_2 > 0 && count_1 == 0) {
-      intensities[j] <- intensities[j] + theta_2[count_2]
+      lambda[j] <- lambda[j] + theta_2[count_2]
     }
   }
 
-  intensities
+  lambda
 }