+++ NEWS.md
@@ -5,6 +5,7 @@
     and arguments/variables `ex_intensities` to `theta`
   - Rename S4-method `uexIntensities()` to `calcExShockArrivalIntensities()`
   - Rename S4-method `exQMatrix()` to `calcMDCMGeneratorMatrix()`
+  - Rename S4-method `intensities()` to `calcShockArrivalIntensities()`
 
 - Improve extensibility by exposing `defaultMethod` and `valueOf0` (#118)